@@ -40,19 +40,19 @@ dependencies:
   flag: ^7.0.0
 
   #Loogs and crash reporting
-  catcher_2: ^1.2.6
-  sentry_flutter: ^8.6.0
+#  catcher_2: ^1.2.6
+  sentry_flutter: ^8.9.0
 
   badges: ^3.1.2
-  dotted_border: ^2.0.0+3
+  dotted_border: ^2.1.0
   styled_text: ^8.1.0
   emoji_picker_flutter: ^1.6.4
   pin_code_text_field: ^1.8.0
   scrollable_positioned_list: ^0.3.8
   infinite_scroll_pagination: ^4.0.0
-  email_validator: ^2.1.17
+  email_validator: ^3.0.0
   credit_card_validator: ^2.1.0
-  flutter_multi_formatter: ^2.12.4
+  flutter_multi_formatter: ^2.13.0
 
   # Media & File handling
   audioplayers: ^6.0.0
@@ -61,21 +61,22 @@ dependencies:
   file_picker: ^8.0.6
   filesize: ^2.0.1
   flutter_image_compress: ^2.3.0
+  fluttertoast: ^8.2.8
 
   # Networking
-  cached_network_image: ^3.3.1
-  dio: ^5.4.3+1
-  internet_connection_checker_plus: ^2.4.2
+  cached_network_image: ^3.4.1
+  dio: ^5.7.0
+  internet_connection_checker_plus: ^2.5.2
 
   # caching
-  flutter_cache_manager: ^3.3.2
+  flutter_cache_manager: ^3.4.1
   # change this with flutter_downloader
   flutter_uploader: ^3.0.0-beta.3
-  mime: ^1.0.4
+  mime: ^1.0.6
   flutter_pdfview: ^1.3.2
 
   # Navigation & Localization
-  auto_route: ^8.2.0
+  auto_route: ^9.2.2
   i18n_extension: ^12.0.1
   i18n_extension_importer: ^0.0.6
   intl: ^0.19.0
@@ -85,7 +86,7 @@ dependencies:
   qr_code_scanner: ^1.0.1
 
   # Timer & Overlay
-  stop_watch_timer: ^3.1.1
+  stop_watch_timer: ^3.2.1
   loader_overlay: ^2.3.2
 
   # Keyboard & Color utilities
@@ -93,8 +94,8 @@ dependencies:
   hexcolor: ^3.0.1
 
   # URL & Sharing utilities
-  url_launcher: ^6.2.5
-  share_plus: ^9.0.0
+  url_launcher: ^6.3.0
+  share_plus: ^10.0.2
   flutter_inappwebview: ^6.0.0
   flutter_windows_webview:
     git:
@@ -102,7 +103,7 @@ dependencies:
       ref: master
 
   # Desktop
-  window_manager: ^0.3.9
+  window_manager: ^0.4.2
   tray_manager: ^0.2.3
   flutter_distributor: ^0.0.2
 
@@ -150,7 +151,7 @@ dev_dependencies:
   flutter_lints: ^4.0.0
   test: ^1.15.7
   mockito: ^5.0.12
-  auto_route_generator: ^8.0.0
+  auto_route_generator: ^9.0.0
   build_runner: ^2.0.5
 
 # For information on the generic Dart part of this file, see the
