@@ -19,7 +19,7 @@ import 'messaging/messaging_model.dart';
 
 @RoutePage(name: 'Home')
 class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
+  const HomePage({super.key});
 
   @override
   _HomePageState createState() => _HomePageState();
