@@ -1,6 +1,6 @@
 import 'package:lantern/messaging/messaging.dart';
 
-@RoutePage<void>(name: 'BlockedUsers')
+@RoutePage(name: 'BlockedUsers')
 class BlockedUsers extends StatelessWidget {
   BlockedUsers({Key? key}) : super(key: key);
 
