@@ -1,6 +1,6 @@
 import 'package:lantern/messaging/messaging.dart';
 
-@RoutePage<void>(name: 'ChatNumberAccount')
+@RoutePage(name: 'ChatNumberAccount')
 class ChatNumberAccount extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
