@@ -3,7 +3,7 @@
 import 'package:email_validator/email_validator.dart';
 import 'package:lantern/common/common.dart';
 
-@RoutePage<void>(name: 'AuthorizeDeviceEmail')
+@RoutePage(name: 'AuthorizeDeviceEmail')
 class AuthorizeDeviceViaEmail extends StatelessWidget {
   AuthorizeDeviceViaEmail({Key? key}) : super(key: key);
 
