@@ -9,9 +9,8 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:io' as _i60;
-import 'dart:ui' as _i58;
 
-import 'package:auto_route/auto_route.dart' as _i55;
+import 'package:auto_route/auto_route.dart' as _i56;
 import 'package:flutter/cupertino.dart' as _i61;
 import 'package:lantern/account/account.dart' as _i2;
 import 'package:lantern/account/account_management.dart' as _i1;
@@ -20,11 +19,11 @@ import 'package:lantern/account/auth/change_email.dart' as _i11;
 import 'package:lantern/account/auth/confirm_email.dart' as _i17;
 import 'package:lantern/account/auth/create_account_email.dart' as _i20;
 import 'package:lantern/account/auth/create_account_password.dart' as _i21;
-import 'package:lantern/account/auth/reset_password.dart' as _i45;
-import 'package:lantern/account/auth/restore_purchase.dart' as _i46;
-import 'package:lantern/account/auth/sign_in.dart' as _i48;
-import 'package:lantern/account/auth/sign_in_password.dart' as _i49;
-import 'package:lantern/account/auth/verification.dart' as _i54;
+import 'package:lantern/account/auth/reset_password.dart' as _i46;
+import 'package:lantern/account/auth/restore_purchase.dart' as _i47;
+import 'package:lantern/account/auth/sign_in.dart' as _i49;
+import 'package:lantern/account/auth/sign_in_password.dart' as _i50;
+import 'package:lantern/account/auth/verification.dart' as _i55;
 import 'package:lantern/account/blocked_users.dart' as _i10;
 import 'package:lantern/account/chat_number_account.dart' as _i12;
 import 'package:lantern/account/device_linking/add_device.dart' as _i3;
@@ -39,12 +38,12 @@ import 'package:lantern/account/device_linking/link_device.dart' as _i30;
 import 'package:lantern/account/invite_friends.dart' as _i27;
 import 'package:lantern/account/language.dart' as _i28;
 import 'package:lantern/account/lantern_desktop.dart' as _i29;
-import 'package:lantern/account/recovery_key.dart' as _i33;
-import 'package:lantern/account/report_issue.dart' as _i42;
-import 'package:lantern/account/settings.dart' as _i47;
-import 'package:lantern/account/split_tunneling.dart' as _i50;
-import 'package:lantern/account/support.dart' as _i53;
-import 'package:lantern/common/common.dart' as _i57;
+import 'package:lantern/account/recovery_key.dart' as _i34;
+import 'package:lantern/account/report_issue.dart' as _i43;
+import 'package:lantern/account/settings.dart' as _i48;
+import 'package:lantern/account/split_tunneling.dart' as _i51;
+import 'package:lantern/account/support.dart' as _i54;
+import 'package:lantern/common/common.dart' as _i58;
 import 'package:lantern/common/ui/app_webview.dart' as _i5;
 import 'package:lantern/common/ui/full_screen_dialog.dart' as _i23;
 import 'package:lantern/home.dart' as _i24;
@@ -54,527 +53,36 @@ import 'package:lantern/messaging/contacts/new_chat.dart' as _i31;
 import 'package:lantern/messaging/conversation/conversation.dart' as _i19;
 import 'package:lantern/messaging/introductions/introduce.dart' as _i25;
 import 'package:lantern/messaging/introductions/introductions.dart' as _i26;
-import 'package:lantern/messaging/messaging.dart' as _i56;
+import 'package:lantern/messaging/messaging.dart' as _i57;
 import 'package:lantern/messaging/onboarding/chat_number_messaging.dart'
     as _i13;
 import 'package:lantern/messaging/onboarding/chat_number_recovery.dart' as _i14;
 import 'package:lantern/plans/checkout.dart' as _i15;
 import 'package:lantern/plans/checkout_legacy.dart' as _i16;
 import 'package:lantern/plans/plans.dart' as _i32;
-import 'package:lantern/plans/reseller_checkout.dart' as _i44;
-import 'package:lantern/plans/reseller_checkout_legacy.dart' as _i43;
-import 'package:lantern/plans/store_checkout.dart' as _i51;
-import 'package:lantern/plans/stripe_checkout.dart' as _i52;
+import 'package:lantern/plans/play_checkout.dart' as _i33;
+import 'package:lantern/plans/reseller_checkout.dart' as _i45;
+import 'package:lantern/plans/reseller_checkout_legacy.dart' as _i44;
+import 'package:lantern/plans/store_checkout.dart' as _i52;
+import 'package:lantern/plans/stripe_checkout.dart' as _i53;
 import 'package:lantern/replica/common.dart' as _i59;
-import 'package:lantern/replica/link_handler.dart' as _i36;
-import 'package:lantern/replica/ui/viewers/audio.dart' as _i34;
-import 'package:lantern/replica/ui/viewers/image.dart' as _i35;
-import 'package:lantern/replica/ui/viewers/misc.dart' as _i37;
-import 'package:lantern/replica/ui/viewers/video.dart' as _i41;
-import 'package:lantern/replica/upload/description.dart' as _i38;
-import 'package:lantern/replica/upload/review.dart' as _i39;
-import 'package:lantern/replica/upload/title.dart' as _i40;
+import 'package:lantern/replica/link_handler.dart' as _i37;
+import 'package:lantern/replica/ui/viewers/audio.dart' as _i35;
+import 'package:lantern/replica/ui/viewers/image.dart' as _i36;
+import 'package:lantern/replica/ui/viewers/misc.dart' as _i38;
+import 'package:lantern/replica/ui/viewers/video.dart' as _i42;
+import 'package:lantern/replica/upload/description.dart' as _i39;
+import 'package:lantern/replica/upload/review.dart' as _i40;
+import 'package:lantern/replica/upload/title.dart' as _i41;
 import 'package:lantern/vpn/vpn.dart' as _i62;
-
-abstract class $AppRouter extends _i55.RootStackRouter {
-  $AppRouter({super.navigatorKey});
-
-  @override
-  final Map<String, _i55.PageFactory> pagesMap = {
-    AccountManagement.name: (routeData) {
-      final args = routeData.argsAs<AccountManagementArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i1.AccountManagement(
-          key: args.key,
-          isPro: args.isPro,
-        ),
-      );
-    },
-    Account.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i2.AccountMenu(),
-      );
-    },
-    ApproveDevice.name: (routeData) {
-      final args = routeData.argsAs<ApproveDeviceArgs>(
-          orElse: () => const ApproveDeviceArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i3.AddDevice(key: args.key),
-      );
-    },
-    AddViaChatNumber.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i4.AddViaChatNumber(),
-      );
-    },
-    AppWebview.name: (routeData) {
-      final args = routeData.argsAs<AppWebviewArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i5.AppWebView(
-          key: args.key,
-          url: args.url,
-          title: args.title,
-        ),
-      );
-    },
-    AuthLanding.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i6.AuthLanding(),
-      );
-    },
-    AuthorizePro.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i7.AuthorizeDeviceForPro(),
-      );
-    },
-    AuthorizeDeviceEmail.name: (routeData) {
-      final args = routeData.argsAs<AuthorizeDeviceEmailArgs>(
-          orElse: () => const AuthorizeDeviceEmailArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i8.AuthorizeDeviceViaEmail(key: args.key),
-      );
-    },
-    AuthorizeDeviceEmailPin.name: (routeData) {
-      final args = routeData.argsAs<AuthorizeDeviceEmailPinArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i9.AuthorizeDeviceViaEmailPin(
-          key: args.key,
-          email: args.email,
-        ),
-      );
-    },
-    BlockedUsers.name: (routeData) {
-      final args = routeData.argsAs<BlockedUsersArgs>(
-          orElse: () => const BlockedUsersArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i10.BlockedUsers(key: args.key),
-      );
-    },
-    ChangeEmail.name: (routeData) {
-      final args = routeData.argsAs<ChangeEmailArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i11.ChangeEmail(
-          key: args.key,
-          email: args.email,
-        ),
-      );
-    },
-    ChatNumberAccount.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i12.ChatNumberAccount(),
-      );
-    },
-    ChatNumberMessaging.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i13.ChatNumberMessaging(),
-      );
-    },
-    ChatNumberRecovery.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i14.ChatNumberRecovery(),
-      );
-    },
-    Checkout.name: (routeData) {
-      final args = routeData.argsAs<CheckoutArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i15.Checkout(
-          plan: args.plan,
-          isPro: args.isPro,
-          authFlow: args.authFlow,
-          email: args.email,
-          verificationPin: args.verificationPin,
-          key: args.key,
-        ),
-      );
-    },
-    CheckoutLegacy.name: (routeData) {
-      final args = routeData.argsAs<CheckoutLegacyArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i16.CheckoutLegacy(
-          plan: args.plan,
-          isPro: args.isPro,
-          key: args.key,
-        ),
-      );
-    },
-    ConfirmEmail.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i17.ConfirmEmail(),
-      );
-    },
-    ContactInfo.name: (routeData) {
-      final args = routeData.argsAs<ContactInfoArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i18.ContactInfo(contact: args.contact),
-      );
-    },
-    Conversation.name: (routeData) {
-      final args = routeData.argsAs<ConversationArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i19.Conversation(
-          contactId: args.contactId,
-          initialScrollIndex: args.initialScrollIndex,
-          showContactEditingDialog: args.showContactEditingDialog,
-        ),
-      );
-    },
-    CreateAccountEmail.name: (routeData) {
-      final args = routeData.argsAs<CreateAccountEmailArgs>(
-          orElse: () => const CreateAccountEmailArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i20.CreateAccountEmail(
-          key: args.key,
-          plan: args.plan,
-          authFlow: args.authFlow,
-          email: args.email,
-        ),
-      );
-    },
-    CreateAccountPassword.name: (routeData) {
-      final args = routeData.argsAs<CreateAccountPasswordArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i21.CreateAccountPassword(
-          key: args.key,
-          email: args.email,
-          code: args.code,
-        ),
-      );
-    },
-    DeviceLimit.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i22.DeviceLimit(),
-      );
-    },
-    FullScreenDialogPage.name: (routeData) {
-      final args = routeData.argsAs<FullScreenDialogPageArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i23.FullScreenDialog(
-          widget: args.widget,
-          bgColor: args.bgColor,
-          key: args.key,
-        ),
-      );
-    },
-    Home.name: (routeData) {
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const _i24.HomePage(),
-      );
-    },
-    Introduce.name: (routeData) {
-      final args = routeData.argsAs<IntroduceArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i25.Introduce(
-          singleIntro: args.singleIntro,
-          contactToIntro: args.contactToIntro,
-        ),
-      );
-    },
-    Introductions.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i26.Introductions(),
-      );
-    },
-    InviteFriends.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i27.InviteFriends(),
-      );
-    },
-    Language.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i28.Language(),
-      );
-    },
-    LanternDesktop.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i29.LanternDesktop(),
-      );
-    },
-    LinkDevice.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i30.LinkDevice(),
-      );
-    },
-    NewChat.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i31.NewChat(),
-      );
-    },
-    PlansPage.name: (routeData) {
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const _i32.PlansPage(),
-      );
-    },
-    RecoveryKey.name: (routeData) {
-      final args = routeData.argsAs<RecoveryKeyArgs>(
-          orElse: () => const RecoveryKeyArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i33.RecoveryKey(key: args.key),
-      );
-    },
-    ReplicaAudioViewer.name: (routeData) {
-      final args = routeData.argsAs<ReplicaAudioViewerArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i34.ReplicaAudioViewer(
-          replicaApi: args.replicaApi,
-          item: args.item,
-          category: args.category,
-        ),
-      );
-    },
-    ReplicaImageViewer.name: (routeData) {
-      final args = routeData.argsAs<ReplicaImageViewerArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i35.ReplicaImageViewer(
-          replicaApi: args.replicaApi,
-          item: args.item,
-          category: args.category,
-        ),
-      );
-    },
-    ReplicaLinkHandler.name: (routeData) {
-      final args = routeData.argsAs<ReplicaLinkHandlerArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i36.ReplicaLinkHandler(
-          key: args.key,
-          replicaApi: args.replicaApi,
-          replicaLink: args.replicaLink,
-        ),
-      );
-    },
-    ReplicaMiscViewer.name: (routeData) {
-      final args = routeData.argsAs<ReplicaMiscViewerArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i37.ReplicaMiscViewer(
-          replicaApi: args.replicaApi,
-          item: args.item,
-          category: args.category,
-        ),
-      );
-    },
-    ReplicaUploadDescription.name: (routeData) {
-      final args = routeData.argsAs<ReplicaUploadDescriptionArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i38.ReplicaUploadDescription(
-          key: args.key,
-          fileToUpload: args.fileToUpload,
-          fileTitle: args.fileTitle,
-          fileDescription: args.fileDescription,
-        ),
-      );
-    },
-    ReplicaUploadReview.name: (routeData) {
-      final args = routeData.argsAs<ReplicaUploadReviewArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i39.ReplicaUploadReview(
-          key: args.key,
-          fileToUpload: args.fileToUpload,
-          fileTitle: args.fileTitle,
-          fileDescription: args.fileDescription,
-        ),
-      );
-    },
-    ReplicaUploadTitle.name: (routeData) {
-      final args = routeData.argsAs<ReplicaUploadTitleArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i40.ReplicaUploadTitle(
-          key: args.key,
-          fileToUpload: args.fileToUpload,
-          fileTitle: args.fileTitle,
-          fileDescription: args.fileDescription,
-        ),
-      );
-    },
-    ReplicaVideoViewer.name: (routeData) {
-      final args = routeData.argsAs<ReplicaVideoViewerArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i41.ReplicaVideoViewer(
-          replicaApi: args.replicaApi,
-          item: args.item,
-          category: args.category,
-        ),
-      );
-    },
-    ReportIssue.name: (routeData) {
-      final args = routeData.argsAs<ReportIssueArgs>(
-          orElse: () => const ReportIssueArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i42.ReportIssue(
-          key: args.key,
-          description: args.description,
-        ),
-      );
-    },
-    ResellerCodeCheckoutLegacy.name: (routeData) {
-      final args = routeData.argsAs<ResellerCodeCheckoutLegacyArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i43.ResellerCodeCheckout(
-          isPro: args.isPro,
-          key: args.key,
-        ),
-      );
-    },
-    ResellerCodeCheckout.name: (routeData) {
-      final args = routeData.argsAs<ResellerCodeCheckoutArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i44.ResellerCodeCheckout(
-          isPro: args.isPro,
-          email: args.email,
-          otp: args.otp,
-          key: args.key,
-        ),
-      );
-    },
-    ResetPassword.name: (routeData) {
-      final args = routeData.argsAs<ResetPasswordArgs>(
-          orElse: () => const ResetPasswordArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i45.ResetPassword(
-          key: args.key,
-          email: args.email,
-          code: args.code,
-          authFlow: args.authFlow,
-        ),
-      );
-    },
-    RestorePurchase.name: (routeData) {
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: const _i46.RestorePurchase(),
-      );
-    },
-    Settings.name: (routeData) {
-      final args =
-          routeData.argsAs<SettingsArgs>(orElse: () => const SettingsArgs());
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i47.Settings(key: args.key),
-      );
-    },
-    SignIn.name: (routeData) {
-      final args =
-          routeData.argsAs<SignInArgs>(orElse: () => const SignInArgs());
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i48.SignIn(
-          key: args.key,
-          authFlow: args.authFlow,
-        ),
-      );
-    },
-    SignInPassword.name: (routeData) {
-      final args = routeData.argsAs<SignInPasswordArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i49.SignInPassword(
-          key: args.key,
-          email: args.email,
-        ),
-      );
-    },
-    SplitTunneling.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i50.SplitTunneling(),
-      );
-    },
-    StoreCheckout.name: (routeData) {
-      final args = routeData.argsAs<StoreCheckoutArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i51.StoreCheckout(
-          plan: args.plan,
-          key: args.key,
-        ),
-      );
-    },
-    StripeCheckout.name: (routeData) {
-      final args = routeData.argsAs<StripeCheckoutArgs>();
-      return _i55.AutoRoutePage<dynamic>(
-        routeData: routeData,
-        child: _i52.StripeCheckout(
-          plan: args.plan,
-          email: args.email,
-          refCode: args.refCode,
-          isPro: args.isPro,
-          key: args.key,
-        ),
-      );
-    },
-    Support.name: (routeData) {
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: const _i53.Support(),
-      );
-    },
-    Verification.name: (routeData) {
-      final args = routeData.argsAs<VerificationArgs>();
-      return _i55.AutoRoutePage<void>(
-        routeData: routeData,
-        child: _i54.Verification(
-          key: args.key,
-          email: args.email,
-          authFlow: args.authFlow,
-          changeEmailArgs: args.changeEmailArgs,
-          plan: args.plan,
-          tempPassword: args.tempPassword,
-        ),
-      );
-    },
-  };
-}
 
 /// generated route for
 /// [_i1.AccountManagement]
-class AccountManagement extends _i55.PageRouteInfo<AccountManagementArgs> {
+class AccountManagement extends _i56.PageRouteInfo<AccountManagementArgs> {
   AccountManagement({
-    _i56.Key? key,
+    _i57.Key? key,
     required bool isPro,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           AccountManagement.name,
           args: AccountManagementArgs(
@@ -586,8 +94,16 @@ class AccountManagement extends _i55.PageRouteInfo<AccountManagementArgs> {
 
   static const String name = 'AccountManagement';
 
-  static const _i55.PageInfo<AccountManagementArgs> page =
-      _i55.PageInfo<AccountManagementArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<AccountManagementArgs>();
+      return _i1.AccountManagement(
+        key: args.key,
+        isPro: args.isPro,
+      );
+    },
+  );
 }
 
 class AccountManagementArgs {
@@ -596,7 +112,7 @@ class AccountManagementArgs {
     required this.isPro,
   });
 
-  final _i56.Key? key;
+  final _i57.Key? key;
 
   final bool isPro;
 
@@ -608,8 +124,8 @@ class AccountManagementArgs {
 
 /// generated route for
 /// [_i2.AccountMenu]
-class Account extends _i55.PageRouteInfo<void> {
-  const Account({List<_i55.PageRouteInfo>? children})
+class Account extends _i56.PageRouteInfo<void> {
+  const Account({List<_i56.PageRouteInfo>? children})
       : super(
           Account.name,
           initialChildren: children,
@@ -617,15 +133,20 @@ class Account extends _i55.PageRouteInfo<void> {
 
   static const String name = 'Account';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i2.AccountMenu();
+    },
+  );
 }
 
 /// generated route for
 /// [_i3.AddDevice]
-class ApproveDevice extends _i55.PageRouteInfo<ApproveDeviceArgs> {
+class ApproveDevice extends _i56.PageRouteInfo<ApproveDeviceArgs> {
   ApproveDevice({
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ApproveDevice.name,
           args: ApproveDeviceArgs(key: key),
@@ -634,14 +155,20 @@ class ApproveDevice extends _i55.PageRouteInfo<ApproveDeviceArgs> {
 
   static const String name = 'ApproveDevice';
 
-  static const _i55.PageInfo<ApproveDeviceArgs> page =
-      _i55.PageInfo<ApproveDeviceArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ApproveDeviceArgs>(
+          orElse: () => const ApproveDeviceArgs());
+      return _i3.AddDevice(key: args.key);
+    },
+  );
 }
 
 class ApproveDeviceArgs {
   const ApproveDeviceArgs({this.key});
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -651,8 +178,8 @@ class ApproveDeviceArgs {
 
 /// generated route for
 /// [_i4.AddViaChatNumber]
-class AddViaChatNumber extends _i55.PageRouteInfo<void> {
-  const AddViaChatNumber({List<_i55.PageRouteInfo>? children})
+class AddViaChatNumber extends _i56.PageRouteInfo<void> {
+  const AddViaChatNumber({List<_i56.PageRouteInfo>? children})
       : super(
           AddViaChatNumber.name,
           initialChildren: children,
@@ -660,17 +187,22 @@ class AddViaChatNumber extends _i55.PageRouteInfo<void> {
 
   static const String name = 'AddViaChatNumber';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i4.AddViaChatNumber();
+    },
+  );
 }
 
 /// generated route for
 /// [_i5.AppWebView]
-class AppWebview extends _i55.PageRouteInfo<AppWebviewArgs> {
+class AppWebview extends _i56.PageRouteInfo<AppWebviewArgs> {
   AppWebview({
-    _i57.Key? key,
+    _i58.Key? key,
     required String url,
     String title = "",
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           AppWebview.name,
           args: AppWebviewArgs(
@@ -683,8 +215,17 @@ class AppWebview extends _i55.PageRouteInfo<AppWebviewArgs> {
 
   static const String name = 'AppWebview';
 
-  static const _i55.PageInfo<AppWebviewArgs> page =
-      _i55.PageInfo<AppWebviewArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<AppWebviewArgs>();
+      return _i5.AppWebView(
+        key: args.key,
+        url: args.url,
+        title: args.title,
+      );
+    },
+  );
 }
 
 class AppWebviewArgs {
@@ -694,7 +235,7 @@ class AppWebviewArgs {
     this.title = "",
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String url;
 
@@ -708,8 +249,8 @@ class AppWebviewArgs {
 
 /// generated route for
 /// [_i6.AuthLanding]
-class AuthLanding extends _i55.PageRouteInfo<void> {
-  const AuthLanding({List<_i55.PageRouteInfo>? children})
+class AuthLanding extends _i56.PageRouteInfo<void> {
+  const AuthLanding({List<_i56.PageRouteInfo>? children})
       : super(
           AuthLanding.name,
           initialChildren: children,
@@ -717,13 +258,18 @@ class AuthLanding extends _i55.PageRouteInfo<void> {
 
   static const String name = 'AuthLanding';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i6.AuthLanding();
+    },
+  );
 }
 
 /// generated route for
 /// [_i7.AuthorizeDeviceForPro]
-class AuthorizePro extends _i55.PageRouteInfo<void> {
-  const AuthorizePro({List<_i55.PageRouteInfo>? children})
+class AuthorizePro extends _i56.PageRouteInfo<void> {
+  const AuthorizePro({List<_i56.PageRouteInfo>? children})
       : super(
           AuthorizePro.name,
           initialChildren: children,
@@ -731,16 +277,21 @@ class AuthorizePro extends _i55.PageRouteInfo<void> {
 
   static const String name = 'AuthorizePro';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i7.AuthorizeDeviceForPro();
+    },
+  );
 }
 
 /// generated route for
 /// [_i8.AuthorizeDeviceViaEmail]
 class AuthorizeDeviceEmail
-    extends _i55.PageRouteInfo<AuthorizeDeviceEmailArgs> {
+    extends _i56.PageRouteInfo<AuthorizeDeviceEmailArgs> {
   AuthorizeDeviceEmail({
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           AuthorizeDeviceEmail.name,
           args: AuthorizeDeviceEmailArgs(key: key),
@@ -749,14 +300,20 @@ class AuthorizeDeviceEmail
 
   static const String name = 'AuthorizeDeviceEmail';
 
-  static const _i55.PageInfo<AuthorizeDeviceEmailArgs> page =
-      _i55.PageInfo<AuthorizeDeviceEmailArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<AuthorizeDeviceEmailArgs>(
+          orElse: () => const AuthorizeDeviceEmailArgs());
+      return _i8.AuthorizeDeviceViaEmail(key: args.key);
+    },
+  );
 }
 
 class AuthorizeDeviceEmailArgs {
   const AuthorizeDeviceEmailArgs({this.key});
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -767,11 +324,11 @@ class AuthorizeDeviceEmailArgs {
 /// generated route for
 /// [_i9.AuthorizeDeviceViaEmailPin]
 class AuthorizeDeviceEmailPin
-    extends _i55.PageRouteInfo<AuthorizeDeviceEmailPinArgs> {
+    extends _i56.PageRouteInfo<AuthorizeDeviceEmailPinArgs> {
   AuthorizeDeviceEmailPin({
-    _i57.Key? key,
+    _i58.Key? key,
     required String email,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           AuthorizeDeviceEmailPin.name,
           args: AuthorizeDeviceEmailPinArgs(
@@ -783,8 +340,16 @@ class AuthorizeDeviceEmailPin
 
   static const String name = 'AuthorizeDeviceEmailPin';
 
-  static const _i55.PageInfo<AuthorizeDeviceEmailPinArgs> page =
-      _i55.PageInfo<AuthorizeDeviceEmailPinArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<AuthorizeDeviceEmailPinArgs>();
+      return _i9.AuthorizeDeviceViaEmailPin(
+        key: args.key,
+        email: args.email,
+      );
+    },
+  );
 }
 
 class AuthorizeDeviceEmailPinArgs {
@@ -793,7 +358,7 @@ class AuthorizeDeviceEmailPinArgs {
     required this.email,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String email;
 
@@ -805,10 +370,10 @@ class AuthorizeDeviceEmailPinArgs {
 
 /// generated route for
 /// [_i10.BlockedUsers]
-class BlockedUsers extends _i55.PageRouteInfo<BlockedUsersArgs> {
+class BlockedUsers extends _i56.PageRouteInfo<BlockedUsersArgs> {
   BlockedUsers({
-    _i56.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i57.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           BlockedUsers.name,
           args: BlockedUsersArgs(key: key),
@@ -817,14 +382,20 @@ class BlockedUsers extends _i55.PageRouteInfo<BlockedUsersArgs> {
 
   static const String name = 'BlockedUsers';
 
-  static const _i55.PageInfo<BlockedUsersArgs> page =
-      _i55.PageInfo<BlockedUsersArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args =
+          data.argsAs<BlockedUsersArgs>(orElse: () => const BlockedUsersArgs());
+      return _i10.BlockedUsers(key: args.key);
+    },
+  );
 }
 
 class BlockedUsersArgs {
   const BlockedUsersArgs({this.key});
 
-  final _i56.Key? key;
+  final _i57.Key? key;
 
   @override
   String toString() {
@@ -834,11 +405,11 @@ class BlockedUsersArgs {
 
 /// generated route for
 /// [_i11.ChangeEmail]
-class ChangeEmail extends _i55.PageRouteInfo<ChangeEmailArgs> {
+class ChangeEmail extends _i56.PageRouteInfo<ChangeEmailArgs> {
   ChangeEmail({
-    _i57.Key? key,
+    _i58.Key? key,
     required String email,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ChangeEmail.name,
           args: ChangeEmailArgs(
@@ -850,8 +421,16 @@ class ChangeEmail extends _i55.PageRouteInfo<ChangeEmailArgs> {
 
   static const String name = 'ChangeEmail';
 
-  static const _i55.PageInfo<ChangeEmailArgs> page =
-      _i55.PageInfo<ChangeEmailArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ChangeEmailArgs>();
+      return _i11.ChangeEmail(
+        key: args.key,
+        email: args.email,
+      );
+    },
+  );
 }
 
 class ChangeEmailArgs {
@@ -860,7 +439,7 @@ class ChangeEmailArgs {
     required this.email,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String email;
 
@@ -872,8 +451,8 @@ class ChangeEmailArgs {
 
 /// generated route for
 /// [_i12.ChatNumberAccount]
-class ChatNumberAccount extends _i55.PageRouteInfo<void> {
-  const ChatNumberAccount({List<_i55.PageRouteInfo>? children})
+class ChatNumberAccount extends _i56.PageRouteInfo<void> {
+  const ChatNumberAccount({List<_i56.PageRouteInfo>? children})
       : super(
           ChatNumberAccount.name,
           initialChildren: children,
@@ -881,13 +460,18 @@ class ChatNumberAccount extends _i55.PageRouteInfo<void> {
 
   static const String name = 'ChatNumberAccount';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i12.ChatNumberAccount();
+    },
+  );
 }
 
 /// generated route for
 /// [_i13.ChatNumberMessaging]
-class ChatNumberMessaging extends _i55.PageRouteInfo<void> {
-  const ChatNumberMessaging({List<_i55.PageRouteInfo>? children})
+class ChatNumberMessaging extends _i56.PageRouteInfo<void> {
+  const ChatNumberMessaging({List<_i56.PageRouteInfo>? children})
       : super(
           ChatNumberMessaging.name,
           initialChildren: children,
@@ -895,13 +479,18 @@ class ChatNumberMessaging extends _i55.PageRouteInfo<void> {
 
   static const String name = 'ChatNumberMessaging';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i13.ChatNumberMessaging();
+    },
+  );
 }
 
 /// generated route for
 /// [_i14.ChatNumberRecovery]
-class ChatNumberRecovery extends _i55.PageRouteInfo<void> {
-  const ChatNumberRecovery({List<_i55.PageRouteInfo>? children})
+class ChatNumberRecovery extends _i56.PageRouteInfo<void> {
+  const ChatNumberRecovery({List<_i56.PageRouteInfo>? children})
       : super(
           ChatNumberRecovery.name,
           initialChildren: children,
@@ -909,20 +498,25 @@ class ChatNumberRecovery extends _i55.PageRouteInfo<void> {
 
   static const String name = 'ChatNumberRecovery';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i14.ChatNumberRecovery();
+    },
+  );
 }
 
 /// generated route for
 /// [_i15.Checkout]
-class Checkout extends _i55.PageRouteInfo<CheckoutArgs> {
+class Checkout extends _i56.PageRouteInfo<CheckoutArgs> {
   Checkout({
-    required _i57.Plan plan,
+    required _i58.Plan plan,
     required bool isPro,
-    _i57.AuthFlow? authFlow,
+    _i58.AuthFlow? authFlow,
     String? email,
     String? verificationPin,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           Checkout.name,
           args: CheckoutArgs(
@@ -938,8 +532,20 @@ class Checkout extends _i55.PageRouteInfo<CheckoutArgs> {
 
   static const String name = 'Checkout';
 
-  static const _i55.PageInfo<CheckoutArgs> page =
-      _i55.PageInfo<CheckoutArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<CheckoutArgs>();
+      return _i15.Checkout(
+        plan: args.plan,
+        isPro: args.isPro,
+        authFlow: args.authFlow,
+        email: args.email,
+        verificationPin: args.verificationPin,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class CheckoutArgs {
@@ -952,17 +558,17 @@ class CheckoutArgs {
     this.key,
   });
 
-  final _i57.Plan plan;
+  final _i58.Plan plan;
 
   final bool isPro;
 
-  final _i57.AuthFlow? authFlow;
+  final _i58.AuthFlow? authFlow;
 
   final String? email;
 
   final String? verificationPin;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -972,12 +578,12 @@ class CheckoutArgs {
 
 /// generated route for
 /// [_i16.CheckoutLegacy]
-class CheckoutLegacy extends _i55.PageRouteInfo<CheckoutLegacyArgs> {
+class CheckoutLegacy extends _i56.PageRouteInfo<CheckoutLegacyArgs> {
   CheckoutLegacy({
-    required _i57.Plan plan,
+    required _i58.Plan plan,
     required bool isPro,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           CheckoutLegacy.name,
           args: CheckoutLegacyArgs(
@@ -990,8 +596,17 @@ class CheckoutLegacy extends _i55.PageRouteInfo<CheckoutLegacyArgs> {
 
   static const String name = 'CheckoutLegacy';
 
-  static const _i55.PageInfo<CheckoutLegacyArgs> page =
-      _i55.PageInfo<CheckoutLegacyArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<CheckoutLegacyArgs>();
+      return _i16.CheckoutLegacy(
+        plan: args.plan,
+        isPro: args.isPro,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class CheckoutLegacyArgs {
@@ -1001,11 +616,11 @@ class CheckoutLegacyArgs {
     this.key,
   });
 
-  final _i57.Plan plan;
+  final _i58.Plan plan;
 
   final bool isPro;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -1015,8 +630,8 @@ class CheckoutLegacyArgs {
 
 /// generated route for
 /// [_i17.ConfirmEmail]
-class ConfirmEmail extends _i55.PageRouteInfo<void> {
-  const ConfirmEmail({List<_i55.PageRouteInfo>? children})
+class ConfirmEmail extends _i56.PageRouteInfo<void> {
+  const ConfirmEmail({List<_i56.PageRouteInfo>? children})
       : super(
           ConfirmEmail.name,
           initialChildren: children,
@@ -1024,15 +639,20 @@ class ConfirmEmail extends _i55.PageRouteInfo<void> {
 
   static const String name = 'ConfirmEmail';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i17.ConfirmEmail();
+    },
+  );
 }
 
 /// generated route for
 /// [_i18.ContactInfo]
-class ContactInfo extends _i55.PageRouteInfo<ContactInfoArgs> {
+class ContactInfo extends _i56.PageRouteInfo<ContactInfoArgs> {
   ContactInfo({
-    required _i56.Contact contact,
-    List<_i55.PageRouteInfo>? children,
+    required _i57.Contact contact,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ContactInfo.name,
           args: ContactInfoArgs(contact: contact),
@@ -1041,14 +661,19 @@ class ContactInfo extends _i55.PageRouteInfo<ContactInfoArgs> {
 
   static const String name = 'ContactInfo';
 
-  static const _i55.PageInfo<ContactInfoArgs> page =
-      _i55.PageInfo<ContactInfoArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ContactInfoArgs>();
+      return _i18.ContactInfo(contact: args.contact);
+    },
+  );
 }
 
 class ContactInfoArgs {
   const ContactInfoArgs({required this.contact});
 
-  final _i56.Contact contact;
+  final _i57.Contact contact;
 
   @override
   String toString() {
@@ -1058,12 +683,12 @@ class ContactInfoArgs {
 
 /// generated route for
 /// [_i19.Conversation]
-class Conversation extends _i55.PageRouteInfo<ConversationArgs> {
+class Conversation extends _i56.PageRouteInfo<ConversationArgs> {
   Conversation({
-    required _i56.ContactId contactId,
+    required _i57.ContactId contactId,
     int? initialScrollIndex,
     bool showContactEditingDialog = false,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           Conversation.name,
           args: ConversationArgs(
@@ -1076,8 +701,17 @@ class Conversation extends _i55.PageRouteInfo<ConversationArgs> {
 
   static const String name = 'Conversation';
 
-  static const _i55.PageInfo<ConversationArgs> page =
-      _i55.PageInfo<ConversationArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ConversationArgs>();
+      return _i19.Conversation(
+        contactId: args.contactId,
+        initialScrollIndex: args.initialScrollIndex,
+        showContactEditingDialog: args.showContactEditingDialog,
+      );
+    },
+  );
 }
 
 class ConversationArgs {
@@ -1087,7 +721,7 @@ class ConversationArgs {
     this.showContactEditingDialog = false,
   });
 
-  final _i56.ContactId contactId;
+  final _i57.ContactId contactId;
 
   final int? initialScrollIndex;
 
@@ -1101,13 +735,13 @@ class ConversationArgs {
 
 /// generated route for
 /// [_i20.CreateAccountEmail]
-class CreateAccountEmail extends _i55.PageRouteInfo<CreateAccountEmailArgs> {
+class CreateAccountEmail extends _i56.PageRouteInfo<CreateAccountEmailArgs> {
   CreateAccountEmail({
-    _i57.Key? key,
-    _i57.Plan? plan,
-    _i57.AuthFlow authFlow = _i57.AuthFlow.createAccount,
+    _i58.Key? key,
+    _i58.Plan? plan,
+    _i58.AuthFlow authFlow = _i58.AuthFlow.createAccount,
     String? email,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           CreateAccountEmail.name,
           args: CreateAccountEmailArgs(
@@ -1121,23 +755,34 @@ class CreateAccountEmail extends _i55.PageRouteInfo<CreateAccountEmailArgs> {
 
   static const String name = 'CreateAccountEmail';
 
-  static const _i55.PageInfo<CreateAccountEmailArgs> page =
-      _i55.PageInfo<CreateAccountEmailArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<CreateAccountEmailArgs>(
+          orElse: () => const CreateAccountEmailArgs());
+      return _i20.CreateAccountEmail(
+        key: args.key,
+        plan: args.plan,
+        authFlow: args.authFlow,
+        email: args.email,
+      );
+    },
+  );
 }
 
 class CreateAccountEmailArgs {
   const CreateAccountEmailArgs({
     this.key,
     this.plan,
-    this.authFlow = _i57.AuthFlow.createAccount,
+    this.authFlow = _i58.AuthFlow.createAccount,
     this.email,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
-  final _i57.Plan? plan;
+  final _i58.Plan? plan;
 
-  final _i57.AuthFlow authFlow;
+  final _i58.AuthFlow authFlow;
 
   final String? email;
 
@@ -1150,12 +795,12 @@ class CreateAccountEmailArgs {
 /// generated route for
 /// [_i21.CreateAccountPassword]
 class CreateAccountPassword
-    extends _i55.PageRouteInfo<CreateAccountPasswordArgs> {
+    extends _i56.PageRouteInfo<CreateAccountPasswordArgs> {
   CreateAccountPassword({
-    _i57.Key? key,
+    _i58.Key? key,
     required String email,
     required String code,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           CreateAccountPassword.name,
           args: CreateAccountPasswordArgs(
@@ -1168,8 +813,17 @@ class CreateAccountPassword
 
   static const String name = 'CreateAccountPassword';
 
-  static const _i55.PageInfo<CreateAccountPasswordArgs> page =
-      _i55.PageInfo<CreateAccountPasswordArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<CreateAccountPasswordArgs>();
+      return _i21.CreateAccountPassword(
+        key: args.key,
+        email: args.email,
+        code: args.code,
+      );
+    },
+  );
 }
 
 class CreateAccountPasswordArgs {
@@ -1179,7 +833,7 @@ class CreateAccountPasswordArgs {
     required this.code,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String email;
 
@@ -1193,8 +847,8 @@ class CreateAccountPasswordArgs {
 
 /// generated route for
 /// [_i22.DeviceLimit]
-class DeviceLimit extends _i55.PageRouteInfo<void> {
-  const DeviceLimit({List<_i55.PageRouteInfo>? children})
+class DeviceLimit extends _i56.PageRouteInfo<void> {
+  const DeviceLimit({List<_i56.PageRouteInfo>? children})
       : super(
           DeviceLimit.name,
           initialChildren: children,
@@ -1202,18 +856,23 @@ class DeviceLimit extends _i55.PageRouteInfo<void> {
 
   static const String name = 'DeviceLimit';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i22.DeviceLimit();
+    },
+  );
 }
 
 /// generated route for
 /// [_i23.FullScreenDialog]
 class FullScreenDialogPage
-    extends _i55.PageRouteInfo<FullScreenDialogPageArgs> {
+    extends _i56.PageRouteInfo<FullScreenDialogPageArgs> {
   FullScreenDialogPage({
-    required _i57.Widget widget,
+    required _i58.Widget widget,
     _i58.Color? bgColor,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           FullScreenDialogPage.name,
           args: FullScreenDialogPageArgs(
@@ -1226,8 +885,17 @@ class FullScreenDialogPage
 
   static const String name = 'FullScreenDialogPage';
 
-  static const _i55.PageInfo<FullScreenDialogPageArgs> page =
-      _i55.PageInfo<FullScreenDialogPageArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<FullScreenDialogPageArgs>();
+      return _i23.FullScreenDialog(
+        widget: args.widget,
+        bgColor: args.bgColor,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class FullScreenDialogPageArgs {
@@ -1237,11 +905,11 @@ class FullScreenDialogPageArgs {
     this.key,
   });
 
-  final _i57.Widget widget;
+  final _i58.Widget widget;
 
   final _i58.Color? bgColor;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -1251,8 +919,8 @@ class FullScreenDialogPageArgs {
 
 /// generated route for
 /// [_i24.HomePage]
-class Home extends _i55.PageRouteInfo<void> {
-  const Home({List<_i55.PageRouteInfo>? children})
+class Home extends _i56.PageRouteInfo<void> {
+  const Home({List<_i56.PageRouteInfo>? children})
       : super(
           Home.name,
           initialChildren: children,
@@ -1260,16 +928,21 @@ class Home extends _i55.PageRouteInfo<void> {
 
   static const String name = 'Home';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i24.HomePage();
+    },
+  );
 }
 
 /// generated route for
 /// [_i25.Introduce]
-class Introduce extends _i55.PageRouteInfo<IntroduceArgs> {
+class Introduce extends _i56.PageRouteInfo<IntroduceArgs> {
   Introduce({
     required bool singleIntro,
-    _i56.Contact? contactToIntro,
-    List<_i55.PageRouteInfo>? children,
+    _i57.Contact? contactToIntro,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           Introduce.name,
           args: IntroduceArgs(
@@ -1281,8 +954,16 @@ class Introduce extends _i55.PageRouteInfo<IntroduceArgs> {
 
   static const String name = 'Introduce';
 
-  static const _i55.PageInfo<IntroduceArgs> page =
-      _i55.PageInfo<IntroduceArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<IntroduceArgs>();
+      return _i25.Introduce(
+        singleIntro: args.singleIntro,
+        contactToIntro: args.contactToIntro,
+      );
+    },
+  );
 }
 
 class IntroduceArgs {
@@ -1293,7 +974,7 @@ class IntroduceArgs {
 
   final bool singleIntro;
 
-  final _i56.Contact? contactToIntro;
+  final _i57.Contact? contactToIntro;
 
   @override
   String toString() {
@@ -1303,8 +984,8 @@ class IntroduceArgs {
 
 /// generated route for
 /// [_i26.Introductions]
-class Introductions extends _i55.PageRouteInfo<void> {
-  const Introductions({List<_i55.PageRouteInfo>? children})
+class Introductions extends _i56.PageRouteInfo<void> {
+  const Introductions({List<_i56.PageRouteInfo>? children})
       : super(
           Introductions.name,
           initialChildren: children,
@@ -1312,13 +993,18 @@ class Introductions extends _i55.PageRouteInfo<void> {
 
   static const String name = 'Introductions';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i26.Introductions();
+    },
+  );
 }
 
 /// generated route for
 /// [_i27.InviteFriends]
-class InviteFriends extends _i55.PageRouteInfo<void> {
-  const InviteFriends({List<_i55.PageRouteInfo>? children})
+class InviteFriends extends _i56.PageRouteInfo<void> {
+  const InviteFriends({List<_i56.PageRouteInfo>? children})
       : super(
           InviteFriends.name,
           initialChildren: children,
@@ -1326,13 +1012,18 @@ class InviteFriends extends _i55.PageRouteInfo<void> {
 
   static const String name = 'InviteFriends';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i27.InviteFriends();
+    },
+  );
 }
 
 /// generated route for
 /// [_i28.Language]
-class Language extends _i55.PageRouteInfo<void> {
-  const Language({List<_i55.PageRouteInfo>? children})
+class Language extends _i56.PageRouteInfo<void> {
+  const Language({List<_i56.PageRouteInfo>? children})
       : super(
           Language.name,
           initialChildren: children,
@@ -1340,13 +1031,18 @@ class Language extends _i55.PageRouteInfo<void> {
 
   static const String name = 'Language';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i28.Language();
+    },
+  );
 }
 
 /// generated route for
 /// [_i29.LanternDesktop]
-class LanternDesktop extends _i55.PageRouteInfo<void> {
-  const LanternDesktop({List<_i55.PageRouteInfo>? children})
+class LanternDesktop extends _i56.PageRouteInfo<void> {
+  const LanternDesktop({List<_i56.PageRouteInfo>? children})
       : super(
           LanternDesktop.name,
           initialChildren: children,
@@ -1354,13 +1050,18 @@ class LanternDesktop extends _i55.PageRouteInfo<void> {
 
   static const String name = 'LanternDesktop';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i29.LanternDesktop();
+    },
+  );
 }
 
 /// generated route for
 /// [_i30.LinkDevice]
-class LinkDevice extends _i55.PageRouteInfo<void> {
-  const LinkDevice({List<_i55.PageRouteInfo>? children})
+class LinkDevice extends _i56.PageRouteInfo<void> {
+  const LinkDevice({List<_i56.PageRouteInfo>? children})
       : super(
           LinkDevice.name,
           initialChildren: children,
@@ -1368,13 +1069,18 @@ class LinkDevice extends _i55.PageRouteInfo<void> {
 
   static const String name = 'LinkDevice';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i30.LinkDevice();
+    },
+  );
 }
 
 /// generated route for
 /// [_i31.NewChat]
-class NewChat extends _i55.PageRouteInfo<void> {
-  const NewChat({List<_i55.PageRouteInfo>? children})
+class NewChat extends _i56.PageRouteInfo<void> {
+  const NewChat({List<_i56.PageRouteInfo>? children})
       : super(
           NewChat.name,
           initialChildren: children,
@@ -1382,13 +1088,18 @@ class NewChat extends _i55.PageRouteInfo<void> {
 
   static const String name = 'NewChat';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return _i31.NewChat();
+    },
+  );
 }
 
 /// generated route for
 /// [_i32.PlansPage]
-class PlansPage extends _i55.PageRouteInfo<void> {
-  const PlansPage({List<_i55.PageRouteInfo>? children})
+class PlansPage extends _i56.PageRouteInfo<void> {
+  const PlansPage({List<_i56.PageRouteInfo>? children})
       : super(
           PlansPage.name,
           initialChildren: children,
@@ -1396,15 +1107,72 @@ class PlansPage extends _i55.PageRouteInfo<void> {
 
   static const String name = 'PlansPage';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i32.PlansPage();
+    },
+  );
 }
 
 /// generated route for
-/// [_i33.RecoveryKey]
-class RecoveryKey extends _i55.PageRouteInfo<RecoveryKeyArgs> {
+/// [_i33.PlayCheckout]
+class PlayCheckout extends _i56.PageRouteInfo<PlayCheckoutArgs> {
+  PlayCheckout({
+    required _i58.Plan plan,
+    required bool isPro,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
+  }) : super(
+          PlayCheckout.name,
+          args: PlayCheckoutArgs(
+            plan: plan,
+            isPro: isPro,
+            key: key,
+          ),
+          initialChildren: children,
+        );
+
+  static const String name = 'PlayCheckout';
+
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<PlayCheckoutArgs>();
+      return _i33.PlayCheckout(
+        plan: args.plan,
+        isPro: args.isPro,
+        key: args.key,
+      );
+    },
+  );
+}
+
+class PlayCheckoutArgs {
+  const PlayCheckoutArgs({
+    required this.plan,
+    required this.isPro,
+    this.key,
+  });
+
+  final _i58.Plan plan;
+
+  final bool isPro;
+
+  final _i58.Key? key;
+
+  @override
+  String toString() {
+    return 'PlayCheckoutArgs{plan: $plan, isPro: $isPro, key: $key}';
+  }
+}
+
+/// generated route for
+/// [_i34.RecoveryKey]
+class RecoveryKey extends _i56.PageRouteInfo<RecoveryKeyArgs> {
   RecoveryKey({
-    _i56.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i57.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           RecoveryKey.name,
           args: RecoveryKeyArgs(key: key),
@@ -1413,14 +1181,20 @@ class RecoveryKey extends _i55.PageRouteInfo<RecoveryKeyArgs> {
 
   static const String name = 'RecoveryKey';
 
-  static const _i55.PageInfo<RecoveryKeyArgs> page =
-      _i55.PageInfo<RecoveryKeyArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args =
+          data.argsAs<RecoveryKeyArgs>(orElse: () => const RecoveryKeyArgs());
+      return _i34.RecoveryKey(key: args.key);
+    },
+  );
 }
 
 class RecoveryKeyArgs {
   const RecoveryKeyArgs({this.key});
 
-  final _i56.Key? key;
+  final _i57.Key? key;
 
   @override
   String toString() {
@@ -1429,13 +1203,13 @@ class RecoveryKeyArgs {
 }
 
 /// generated route for
-/// [_i34.ReplicaAudioViewer]
-class ReplicaAudioViewer extends _i55.PageRouteInfo<ReplicaAudioViewerArgs> {
+/// [_i35.ReplicaAudioViewer]
+class ReplicaAudioViewer extends _i56.PageRouteInfo<ReplicaAudioViewerArgs> {
   ReplicaAudioViewer({
     required _i59.ReplicaApi replicaApi,
     required _i59.ReplicaSearchItem item,
     required _i59.SearchCategory category,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaAudioViewer.name,
           args: ReplicaAudioViewerArgs(
@@ -1448,8 +1222,17 @@ class ReplicaAudioViewer extends _i55.PageRouteInfo<ReplicaAudioViewerArgs> {
 
   static const String name = 'ReplicaAudioViewer';
 
-  static const _i55.PageInfo<ReplicaAudioViewerArgs> page =
-      _i55.PageInfo<ReplicaAudioViewerArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaAudioViewerArgs>();
+      return _i35.ReplicaAudioViewer(
+        replicaApi: args.replicaApi,
+        item: args.item,
+        category: args.category,
+      );
+    },
+  );
 }
 
 class ReplicaAudioViewerArgs {
@@ -1472,13 +1255,13 @@ class ReplicaAudioViewerArgs {
 }
 
 /// generated route for
-/// [_i35.ReplicaImageViewer]
-class ReplicaImageViewer extends _i55.PageRouteInfo<ReplicaImageViewerArgs> {
+/// [_i36.ReplicaImageViewer]
+class ReplicaImageViewer extends _i56.PageRouteInfo<ReplicaImageViewerArgs> {
   ReplicaImageViewer({
     required _i59.ReplicaApi replicaApi,
     required _i59.ReplicaSearchItem item,
     required _i59.SearchCategory category,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaImageViewer.name,
           args: ReplicaImageViewerArgs(
@@ -1491,8 +1274,17 @@ class ReplicaImageViewer extends _i55.PageRouteInfo<ReplicaImageViewerArgs> {
 
   static const String name = 'ReplicaImageViewer';
 
-  static const _i55.PageInfo<ReplicaImageViewerArgs> page =
-      _i55.PageInfo<ReplicaImageViewerArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaImageViewerArgs>();
+      return _i36.ReplicaImageViewer(
+        replicaApi: args.replicaApi,
+        item: args.item,
+        category: args.category,
+      );
+    },
+  );
 }
 
 class ReplicaImageViewerArgs {
@@ -1515,13 +1307,13 @@ class ReplicaImageViewerArgs {
 }
 
 /// generated route for
-/// [_i36.ReplicaLinkHandler]
-class ReplicaLinkHandler extends _i55.PageRouteInfo<ReplicaLinkHandlerArgs> {
+/// [_i37.ReplicaLinkHandler]
+class ReplicaLinkHandler extends _i56.PageRouteInfo<ReplicaLinkHandlerArgs> {
   ReplicaLinkHandler({
-    _i57.Key? key,
+    _i58.Key? key,
     required _i59.ReplicaApi replicaApi,
     required _i59.ReplicaLink replicaLink,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaLinkHandler.name,
           args: ReplicaLinkHandlerArgs(
@@ -1534,8 +1326,17 @@ class ReplicaLinkHandler extends _i55.PageRouteInfo<ReplicaLinkHandlerArgs> {
 
   static const String name = 'ReplicaLinkHandler';
 
-  static const _i55.PageInfo<ReplicaLinkHandlerArgs> page =
-      _i55.PageInfo<ReplicaLinkHandlerArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaLinkHandlerArgs>();
+      return _i37.ReplicaLinkHandler(
+        key: args.key,
+        replicaApi: args.replicaApi,
+        replicaLink: args.replicaLink,
+      );
+    },
+  );
 }
 
 class ReplicaLinkHandlerArgs {
@@ -1545,7 +1346,7 @@ class ReplicaLinkHandlerArgs {
     required this.replicaLink,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final _i59.ReplicaApi replicaApi;
 
@@ -1558,13 +1359,13 @@ class ReplicaLinkHandlerArgs {
 }
 
 /// generated route for
-/// [_i37.ReplicaMiscViewer]
-class ReplicaMiscViewer extends _i55.PageRouteInfo<ReplicaMiscViewerArgs> {
+/// [_i38.ReplicaMiscViewer]
+class ReplicaMiscViewer extends _i56.PageRouteInfo<ReplicaMiscViewerArgs> {
   ReplicaMiscViewer({
     required _i59.ReplicaApi replicaApi,
     required _i59.ReplicaSearchItem item,
     required _i59.SearchCategory category,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaMiscViewer.name,
           args: ReplicaMiscViewerArgs(
@@ -1577,8 +1378,17 @@ class ReplicaMiscViewer extends _i55.PageRouteInfo<ReplicaMiscViewerArgs> {
 
   static const String name = 'ReplicaMiscViewer';
 
-  static const _i55.PageInfo<ReplicaMiscViewerArgs> page =
-      _i55.PageInfo<ReplicaMiscViewerArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaMiscViewerArgs>();
+      return _i38.ReplicaMiscViewer(
+        replicaApi: args.replicaApi,
+        item: args.item,
+        category: args.category,
+      );
+    },
+  );
 }
 
 class ReplicaMiscViewerArgs {
@@ -1601,15 +1411,15 @@ class ReplicaMiscViewerArgs {
 }
 
 /// generated route for
-/// [_i38.ReplicaUploadDescription]
+/// [_i39.ReplicaUploadDescription]
 class ReplicaUploadDescription
-    extends _i55.PageRouteInfo<ReplicaUploadDescriptionArgs> {
+    extends _i56.PageRouteInfo<ReplicaUploadDescriptionArgs> {
   ReplicaUploadDescription({
-    _i57.Key? key,
+    _i58.Key? key,
     required _i60.File fileToUpload,
     required String fileTitle,
     String? fileDescription,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaUploadDescription.name,
           args: ReplicaUploadDescriptionArgs(
@@ -1623,8 +1433,18 @@ class ReplicaUploadDescription
 
   static const String name = 'ReplicaUploadDescription';
 
-  static const _i55.PageInfo<ReplicaUploadDescriptionArgs> page =
-      _i55.PageInfo<ReplicaUploadDescriptionArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaUploadDescriptionArgs>();
+      return _i39.ReplicaUploadDescription(
+        key: args.key,
+        fileToUpload: args.fileToUpload,
+        fileTitle: args.fileTitle,
+        fileDescription: args.fileDescription,
+      );
+    },
+  );
 }
 
 class ReplicaUploadDescriptionArgs {
@@ -1635,7 +1455,7 @@ class ReplicaUploadDescriptionArgs {
     this.fileDescription,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final _i60.File fileToUpload;
 
@@ -1650,14 +1470,14 @@ class ReplicaUploadDescriptionArgs {
 }
 
 /// generated route for
-/// [_i39.ReplicaUploadReview]
-class ReplicaUploadReview extends _i55.PageRouteInfo<ReplicaUploadReviewArgs> {
+/// [_i40.ReplicaUploadReview]
+class ReplicaUploadReview extends _i56.PageRouteInfo<ReplicaUploadReviewArgs> {
   ReplicaUploadReview({
-    _i57.Key? key,
+    _i58.Key? key,
     required _i60.File fileToUpload,
     required String fileTitle,
     String? fileDescription,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaUploadReview.name,
           args: ReplicaUploadReviewArgs(
@@ -1671,8 +1491,18 @@ class ReplicaUploadReview extends _i55.PageRouteInfo<ReplicaUploadReviewArgs> {
 
   static const String name = 'ReplicaUploadReview';
 
-  static const _i55.PageInfo<ReplicaUploadReviewArgs> page =
-      _i55.PageInfo<ReplicaUploadReviewArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaUploadReviewArgs>();
+      return _i40.ReplicaUploadReview(
+        key: args.key,
+        fileToUpload: args.fileToUpload,
+        fileTitle: args.fileTitle,
+        fileDescription: args.fileDescription,
+      );
+    },
+  );
 }
 
 class ReplicaUploadReviewArgs {
@@ -1683,7 +1513,7 @@ class ReplicaUploadReviewArgs {
     this.fileDescription,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final _i60.File fileToUpload;
 
@@ -1698,14 +1528,14 @@ class ReplicaUploadReviewArgs {
 }
 
 /// generated route for
-/// [_i40.ReplicaUploadTitle]
-class ReplicaUploadTitle extends _i55.PageRouteInfo<ReplicaUploadTitleArgs> {
+/// [_i41.ReplicaUploadTitle]
+class ReplicaUploadTitle extends _i56.PageRouteInfo<ReplicaUploadTitleArgs> {
   ReplicaUploadTitle({
-    _i57.Key? key,
+    _i58.Key? key,
     required _i60.File fileToUpload,
     String? fileTitle,
     String? fileDescription,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaUploadTitle.name,
           args: ReplicaUploadTitleArgs(
@@ -1719,8 +1549,18 @@ class ReplicaUploadTitle extends _i55.PageRouteInfo<ReplicaUploadTitleArgs> {
 
   static const String name = 'ReplicaUploadTitle';
 
-  static const _i55.PageInfo<ReplicaUploadTitleArgs> page =
-      _i55.PageInfo<ReplicaUploadTitleArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaUploadTitleArgs>();
+      return _i41.ReplicaUploadTitle(
+        key: args.key,
+        fileToUpload: args.fileToUpload,
+        fileTitle: args.fileTitle,
+        fileDescription: args.fileDescription,
+      );
+    },
+  );
 }
 
 class ReplicaUploadTitleArgs {
@@ -1731,7 +1571,7 @@ class ReplicaUploadTitleArgs {
     this.fileDescription,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final _i60.File fileToUpload;
 
@@ -1746,13 +1586,13 @@ class ReplicaUploadTitleArgs {
 }
 
 /// generated route for
-/// [_i41.ReplicaVideoViewer]
-class ReplicaVideoViewer extends _i55.PageRouteInfo<ReplicaVideoViewerArgs> {
+/// [_i42.ReplicaVideoViewer]
+class ReplicaVideoViewer extends _i56.PageRouteInfo<ReplicaVideoViewerArgs> {
   ReplicaVideoViewer({
     required _i59.ReplicaApi replicaApi,
     required _i59.ReplicaSearchItem item,
     required _i59.SearchCategory category,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReplicaVideoViewer.name,
           args: ReplicaVideoViewerArgs(
@@ -1765,8 +1605,17 @@ class ReplicaVideoViewer extends _i55.PageRouteInfo<ReplicaVideoViewerArgs> {
 
   static const String name = 'ReplicaVideoViewer';
 
-  static const _i55.PageInfo<ReplicaVideoViewerArgs> page =
-      _i55.PageInfo<ReplicaVideoViewerArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ReplicaVideoViewerArgs>();
+      return _i42.ReplicaVideoViewer(
+        replicaApi: args.replicaApi,
+        item: args.item,
+        category: args.category,
+      );
+    },
+  );
 }
 
 class ReplicaVideoViewerArgs {
@@ -1789,12 +1638,12 @@ class ReplicaVideoViewerArgs {
 }
 
 /// generated route for
-/// [_i42.ReportIssue]
-class ReportIssue extends _i55.PageRouteInfo<ReportIssueArgs> {
+/// [_i43.ReportIssue]
+class ReportIssue extends _i56.PageRouteInfo<ReportIssueArgs> {
   ReportIssue({
-    _i57.Key? key,
+    _i58.Key? key,
     String? description,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ReportIssue.name,
           args: ReportIssueArgs(
@@ -1806,8 +1655,17 @@ class ReportIssue extends _i55.PageRouteInfo<ReportIssueArgs> {
 
   static const String name = 'ReportIssue';
 
-  static const _i55.PageInfo<ReportIssueArgs> page =
-      _i55.PageInfo<ReportIssueArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args =
+          data.argsAs<ReportIssueArgs>(orElse: () => const ReportIssueArgs());
+      return _i43.ReportIssue(
+        key: args.key,
+        description: args.description,
+      );
+    },
+  );
 }
 
 class ReportIssueArgs {
@@ -1816,7 +1674,7 @@ class ReportIssueArgs {
     this.description,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String? description;
 
@@ -1827,13 +1685,13 @@ class ReportIssueArgs {
 }
 
 /// generated route for
-/// [_i43.ResellerCodeCheckout]
+/// [_i44.ResellerCodeCheckout]
 class ResellerCodeCheckoutLegacy
-    extends _i55.PageRouteInfo<ResellerCodeCheckoutLegacyArgs> {
+    extends _i56.PageRouteInfo<ResellerCodeCheckoutLegacyArgs> {
   ResellerCodeCheckoutLegacy({
     required bool isPro,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ResellerCodeCheckoutLegacy.name,
           args: ResellerCodeCheckoutLegacyArgs(
@@ -1845,8 +1703,16 @@ class ResellerCodeCheckoutLegacy
 
   static const String name = 'ResellerCodeCheckoutLegacy';
 
-  static const _i55.PageInfo<ResellerCodeCheckoutLegacyArgs> page =
-      _i55.PageInfo<ResellerCodeCheckoutLegacyArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ResellerCodeCheckoutLegacyArgs>();
+      return _i44.ResellerCodeCheckout(
+        isPro: args.isPro,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class ResellerCodeCheckoutLegacyArgs {
@@ -1857,7 +1723,7 @@ class ResellerCodeCheckoutLegacyArgs {
 
   final bool isPro;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -1866,15 +1732,15 @@ class ResellerCodeCheckoutLegacyArgs {
 }
 
 /// generated route for
-/// [_i44.ResellerCodeCheckout]
+/// [_i45.ResellerCodeCheckout]
 class ResellerCodeCheckout
-    extends _i55.PageRouteInfo<ResellerCodeCheckoutArgs> {
+    extends _i56.PageRouteInfo<ResellerCodeCheckoutArgs> {
   ResellerCodeCheckout({
     required bool isPro,
     required String email,
     String? otp,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ResellerCodeCheckout.name,
           args: ResellerCodeCheckoutArgs(
@@ -1888,8 +1754,18 @@ class ResellerCodeCheckout
 
   static const String name = 'ResellerCodeCheckout';
 
-  static const _i55.PageInfo<ResellerCodeCheckoutArgs> page =
-      _i55.PageInfo<ResellerCodeCheckoutArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ResellerCodeCheckoutArgs>();
+      return _i45.ResellerCodeCheckout(
+        isPro: args.isPro,
+        email: args.email,
+        otp: args.otp,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class ResellerCodeCheckoutArgs {
@@ -1906,7 +1782,7 @@ class ResellerCodeCheckoutArgs {
 
   final String? otp;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -1915,14 +1791,14 @@ class ResellerCodeCheckoutArgs {
 }
 
 /// generated route for
-/// [_i45.ResetPassword]
-class ResetPassword extends _i55.PageRouteInfo<ResetPasswordArgs> {
+/// [_i46.ResetPassword]
+class ResetPassword extends _i56.PageRouteInfo<ResetPasswordArgs> {
   ResetPassword({
     _i61.Key? key,
     String? email,
     String? code,
-    _i57.AuthFlow authFlow = _i57.AuthFlow.reset,
-    List<_i55.PageRouteInfo>? children,
+    _i58.AuthFlow authFlow = _i58.AuthFlow.reset,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           ResetPassword.name,
           args: ResetPasswordArgs(
@@ -1936,8 +1812,19 @@ class ResetPassword extends _i55.PageRouteInfo<ResetPasswordArgs> {
 
   static const String name = 'ResetPassword';
 
-  static const _i55.PageInfo<ResetPasswordArgs> page =
-      _i55.PageInfo<ResetPasswordArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<ResetPasswordArgs>(
+          orElse: () => const ResetPasswordArgs());
+      return _i46.ResetPassword(
+        key: args.key,
+        email: args.email,
+        code: args.code,
+        authFlow: args.authFlow,
+      );
+    },
+  );
 }
 
 class ResetPasswordArgs {
@@ -1945,7 +1832,7 @@ class ResetPasswordArgs {
     this.key,
     this.email,
     this.code,
-    this.authFlow = _i57.AuthFlow.reset,
+    this.authFlow = _i58.AuthFlow.reset,
   });
 
   final _i61.Key? key;
@@ -1954,7 +1841,7 @@ class ResetPasswordArgs {
 
   final String? code;
 
-  final _i57.AuthFlow authFlow;
+  final _i58.AuthFlow authFlow;
 
   @override
   String toString() {
@@ -1963,9 +1850,9 @@ class ResetPasswordArgs {
 }
 
 /// generated route for
-/// [_i46.RestorePurchase]
-class RestorePurchase extends _i55.PageRouteInfo<void> {
-  const RestorePurchase({List<_i55.PageRouteInfo>? children})
+/// [_i47.RestorePurchase]
+class RestorePurchase extends _i56.PageRouteInfo<void> {
+  const RestorePurchase({List<_i56.PageRouteInfo>? children})
       : super(
           RestorePurchase.name,
           initialChildren: children,
@@ -1973,15 +1860,20 @@ class RestorePurchase extends _i55.PageRouteInfo<void> {
 
   static const String name = 'RestorePurchase';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i47.RestorePurchase();
+    },
+  );
 }
 
 /// generated route for
-/// [_i47.Settings]
-class Settings extends _i55.PageRouteInfo<SettingsArgs> {
+/// [_i48.Settings]
+class Settings extends _i56.PageRouteInfo<SettingsArgs> {
   Settings({
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           Settings.name,
           args: SettingsArgs(key: key),
@@ -1990,14 +1882,20 @@ class Settings extends _i55.PageRouteInfo<SettingsArgs> {
 
   static const String name = 'Settings';
 
-  static const _i55.PageInfo<SettingsArgs> page =
-      _i55.PageInfo<SettingsArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args =
+          data.argsAs<SettingsArgs>(orElse: () => const SettingsArgs());
+      return _i48.Settings(key: args.key);
+    },
+  );
 }
 
 class SettingsArgs {
   const SettingsArgs({this.key});
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -2006,12 +1904,12 @@ class SettingsArgs {
 }
 
 /// generated route for
-/// [_i48.SignIn]
-class SignIn extends _i55.PageRouteInfo<SignInArgs> {
+/// [_i49.SignIn]
+class SignIn extends _i56.PageRouteInfo<SignInArgs> {
   SignIn({
-    _i57.Key? key,
-    _i57.AuthFlow authFlow = _i57.AuthFlow.signIn,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    _i58.AuthFlow authFlow = _i58.AuthFlow.signIn,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           SignIn.name,
           args: SignInArgs(
@@ -2023,18 +1921,27 @@ class SignIn extends _i55.PageRouteInfo<SignInArgs> {
 
   static const String name = 'SignIn';
 
-  static const _i55.PageInfo<SignInArgs> page = _i55.PageInfo<SignInArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<SignInArgs>(orElse: () => const SignInArgs());
+      return _i49.SignIn(
+        key: args.key,
+        authFlow: args.authFlow,
+      );
+    },
+  );
 }
 
 class SignInArgs {
   const SignInArgs({
     this.key,
-    this.authFlow = _i57.AuthFlow.signIn,
+    this.authFlow = _i58.AuthFlow.signIn,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
-  final _i57.AuthFlow authFlow;
+  final _i58.AuthFlow authFlow;
 
   @override
   String toString() {
@@ -2043,12 +1950,12 @@ class SignInArgs {
 }
 
 /// generated route for
-/// [_i49.SignInPassword]
-class SignInPassword extends _i55.PageRouteInfo<SignInPasswordArgs> {
+/// [_i50.SignInPassword]
+class SignInPassword extends _i56.PageRouteInfo<SignInPasswordArgs> {
   SignInPassword({
-    _i57.Key? key,
+    _i58.Key? key,
     required String email,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           SignInPassword.name,
           args: SignInPasswordArgs(
@@ -2060,8 +1967,16 @@ class SignInPassword extends _i55.PageRouteInfo<SignInPasswordArgs> {
 
   static const String name = 'SignInPassword';
 
-  static const _i55.PageInfo<SignInPasswordArgs> page =
-      _i55.PageInfo<SignInPasswordArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<SignInPasswordArgs>();
+      return _i50.SignInPassword(
+        key: args.key,
+        email: args.email,
+      );
+    },
+  );
 }
 
 class SignInPasswordArgs {
@@ -2070,7 +1985,7 @@ class SignInPasswordArgs {
     required this.email,
   });
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   final String email;
 
@@ -2081,9 +1996,9 @@ class SignInPasswordArgs {
 }
 
 /// generated route for
-/// [_i50.SplitTunneling]
-class SplitTunneling extends _i55.PageRouteInfo<void> {
-  const SplitTunneling({List<_i55.PageRouteInfo>? children})
+/// [_i51.SplitTunneling]
+class SplitTunneling extends _i56.PageRouteInfo<void> {
+  const SplitTunneling({List<_i56.PageRouteInfo>? children})
       : super(
           SplitTunneling.name,
           initialChildren: children,
@@ -2091,16 +2006,21 @@ class SplitTunneling extends _i55.PageRouteInfo<void> {
 
   static const String name = 'SplitTunneling';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i51.SplitTunneling();
+    },
+  );
 }
 
 /// generated route for
-/// [_i51.StoreCheckout]
-class StoreCheckout extends _i55.PageRouteInfo<StoreCheckoutArgs> {
+/// [_i52.StoreCheckout]
+class StoreCheckout extends _i56.PageRouteInfo<StoreCheckoutArgs> {
   StoreCheckout({
-    required _i57.Plan plan,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    required _i58.Plan plan,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           StoreCheckout.name,
           args: StoreCheckoutArgs(
@@ -2112,8 +2032,16 @@ class StoreCheckout extends _i55.PageRouteInfo<StoreCheckoutArgs> {
 
   static const String name = 'StoreCheckout';
 
-  static const _i55.PageInfo<StoreCheckoutArgs> page =
-      _i55.PageInfo<StoreCheckoutArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<StoreCheckoutArgs>();
+      return _i52.StoreCheckout(
+        plan: args.plan,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class StoreCheckoutArgs {
@@ -2122,9 +2050,9 @@ class StoreCheckoutArgs {
     this.key,
   });
 
-  final _i57.Plan plan;
+  final _i58.Plan plan;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -2133,15 +2061,15 @@ class StoreCheckoutArgs {
 }
 
 /// generated route for
-/// [_i52.StripeCheckout]
-class StripeCheckout extends _i55.PageRouteInfo<StripeCheckoutArgs> {
+/// [_i53.StripeCheckout]
+class StripeCheckout extends _i56.PageRouteInfo<StripeCheckoutArgs> {
   StripeCheckout({
-    required _i57.Plan plan,
+    required _i58.Plan plan,
     required String email,
     String? refCode,
     required bool isPro,
-    _i57.Key? key,
-    List<_i55.PageRouteInfo>? children,
+    _i58.Key? key,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           StripeCheckout.name,
           args: StripeCheckoutArgs(
@@ -2156,8 +2084,19 @@ class StripeCheckout extends _i55.PageRouteInfo<StripeCheckoutArgs> {
 
   static const String name = 'StripeCheckout';
 
-  static const _i55.PageInfo<StripeCheckoutArgs> page =
-      _i55.PageInfo<StripeCheckoutArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<StripeCheckoutArgs>();
+      return _i53.StripeCheckout(
+        plan: args.plan,
+        email: args.email,
+        refCode: args.refCode,
+        isPro: args.isPro,
+        key: args.key,
+      );
+    },
+  );
 }
 
 class StripeCheckoutArgs {
@@ -2169,7 +2108,7 @@ class StripeCheckoutArgs {
     this.key,
   });
 
-  final _i57.Plan plan;
+  final _i58.Plan plan;
 
   final String email;
 
@@ -2177,7 +2116,7 @@ class StripeCheckoutArgs {
 
   final bool isPro;
 
-  final _i57.Key? key;
+  final _i58.Key? key;
 
   @override
   String toString() {
@@ -2186,9 +2125,9 @@ class StripeCheckoutArgs {
 }
 
 /// generated route for
-/// [_i53.Support]
-class Support extends _i55.PageRouteInfo<void> {
-  const Support({List<_i55.PageRouteInfo>? children})
+/// [_i54.Support]
+class Support extends _i56.PageRouteInfo<void> {
+  const Support({List<_i56.PageRouteInfo>? children})
       : super(
           Support.name,
           initialChildren: children,
@@ -2196,12 +2135,17 @@ class Support extends _i55.PageRouteInfo<void> {
 
   static const String name = 'Support';
 
-  static const _i55.PageInfo<void> page = _i55.PageInfo<void>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      return const _i54.Support();
+    },
+  );
 }
 
 /// generated route for
-/// [_i54.Verification]
-class Verification extends _i55.PageRouteInfo<VerificationArgs> {
+/// [_i55.Verification]
+class Verification extends _i56.PageRouteInfo<VerificationArgs> {
   Verification({
     _i62.Key? key,
     required String email,
@@ -2209,7 +2153,7 @@ class Verification extends _i55.PageRouteInfo<VerificationArgs> {
     _i11.ChangeEmailPageArgs? changeEmailArgs,
     _i62.Plan? plan,
     String? tempPassword,
-    List<_i55.PageRouteInfo>? children,
+    List<_i56.PageRouteInfo>? children,
   }) : super(
           Verification.name,
           args: VerificationArgs(
@@ -2225,8 +2169,20 @@ class Verification extends _i55.PageRouteInfo<VerificationArgs> {
 
   static const String name = 'Verification';
 
-  static const _i55.PageInfo<VerificationArgs> page =
-      _i55.PageInfo<VerificationArgs>(name);
+  static _i56.PageInfo page = _i56.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<VerificationArgs>();
+      return _i55.Verification(
+        key: args.key,
+        email: args.email,
+        authFlow: args.authFlow,
+        changeEmailArgs: args.changeEmailArgs,
+        plan: args.plan,
+        tempPassword: args.tempPassword,
+      );
+    },
+  );
 }
 
 class VerificationArgs {
