@@ -1,6 +1,6 @@
 import 'package:lantern/messaging/messaging.dart';
 
-@RoutePage<void>(name: 'AddViaChatNumber')
+@RoutePage(name: 'AddViaChatNumber')
 class AddViaChatNumber extends StatefulWidget {
   @override
   _AddViaChatNumberState createState() => _AddViaChatNumberState();
