@@ -2,7 +2,7 @@ import 'package:lantern/messaging/conversation/call_action.dart';
 
 import '../messaging.dart';
 
-@RoutePage<void>(name: 'ContactInfo')
+@RoutePage(name: 'ContactInfo')
 class ContactInfo extends StatefulWidget {
   final Contact contact;
 
