@@ -20,7 +20,7 @@ import 'package:lantern/replica/common.dart';
 /// tab will not be visible if Replica is not initialized.
 /// In other words, we will never reach any Replica screen (other than this) if
 /// Replica is not initialized.
-@RoutePage<void>(name: 'ReplicaLinkHandler')
+@RoutePage(name: 'ReplicaLinkHandler')
 class ReplicaLinkHandler extends StatefulWidget {
   ReplicaLinkHandler({
     Key? key,
