@@ -1,6 +1,6 @@
 import 'package:lantern/common/common.dart';
 
-@RoutePage<void>(name: 'ReplicaUploadDescription')
+@RoutePage(name: 'ReplicaUploadDescription')
 class ReplicaUploadDescription extends StatefulWidget {
   final File fileToUpload;
   final String fileTitle;
