@@ -1,4 +1,4 @@
 export 'package:lantern/common/common.dart';
-export 'package:lantern/vpn/protos_shared/vpn.pb.dart';
+
 
 export 'vpn_model.dart';
